django web application/youtubeproject/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
Brotli==1.0.9
channels==2.4.0
channels-redis==4.0.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.3
daphne==2.5.0
dash==2.6.2
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
Django==3.2.16
django-bootstrap4==22.2
django-plotly-dash==2.1.0
django-redis==5.2.0
dpd-components==0.1.0
dpd-static-support==0.0.5
Flask==2.1.3
Flask-Compress==1.13
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
itsdangerous==2.1.2
joblib==1.2.0
mistune==0.8.4
msgpack==1.0.4
numpy==1.23.4
pandas==1.5.1
plotly==5.11.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyOpenSSL==22.1.0
python-decouple==3.6
pytz==2022.6
redis==4.3.4
scikit-learn==0.23.2
scipy==1.9.3
service-identity==21.1.0
sqlparse==0.4.3
tenacity==8.1.0
threadpoolctl==3.1.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.2.0
wincertstore==0.2
wrapt==1.14.1
zipp==3.10.0
zope.interface==5.5.0
